--- token_type/exporter.py
@@ -1,3 +1,0 @@
-# pylint: disable=unused-import
-from token_type.token import Token
-from token_type.token_type_enum import TokenType
